requirements.txt
pandas
numpy
faiss-cpu
torch
sentence-transformers
scikit-learn
umap-learn
matplotlib
seaborn
plotly
wordcloud
fastapi
uvicorn
jinja2